LectureSupplements/MoarCats/requirements.txt
Pillow==2.6.1
numpy==1.9.1
scikit-learn==0.15.2
scipy==0.14.0
wsgiref==0.1.2
